Convert delete hook to POST #2033

+++ modules/Core/pages/panel/hooks.php
@@ -213,20 +213,25 @@ if (!isset($_GET['action'])) {
             break;
 
         case 'delete':
-            // Delete Form
+            // Delete hook
             if (!isset($_GET['id']) || !is_numeric($_GET['id'])) {
                 Redirect::to(URL::build('/panel/core/hooks'));
                 die();
             }
 
-            $queries->delete('hooks', array('id', '=', $_GET['id']));
+            if (Token::check()) {
+                $queries->delete('hooks', array('id', '=', $_GET['id']));
 
-            $cache->setCache('hooks');
-            if ($cache->isCached('hooks')) {
-                $cache->erase('hooks');
+                $cache->setCache('hooks');
+                if ($cache->isCached('hooks')) {
+                    $cache->erase('hooks');
+                }
+
+                Session::flash('admin_hooks', $language->get('admin', 'hook_deleted'));
+            } else {
+                Session::flash('admin_hooks_error', $language->get('general', 'invalid_token'));
             }
 
-            Session::flash('admin_hooks', $language->get('admin', 'hook_deleted'));
             Redirect::to(URL::build('/panel/core/hooks'));
             die();
             break;
@@ -244,8 +249,8 @@ Module::loadPage($user, $pages, $cache, $smarty, array($navigation, $cc_nav, $st
 if (Session::exists('admin_hooks'))
     $success = Session::flash('admin_hooks');
 
-if (Session::exists('admin_pages_error'))
-    $errors[] = Session::flash('admin_pages_error');
+if (Session::exists('admin_hooks_error'))
+    $errors[] = Session::flash('admin_hooks_error');
 
 if (isset($success))
     $smarty->assign(array(
